Do not reload carousel on scroll down on mobile devices
Address bar is hidden on scroll down on most mobile devices, which
triggers a resize event. However Siema does not rely on height, hence
the resize event should be ignored to avoid recreating the DOM for nothing.

This is important for any carousel that relies on DOM construction (ex: videos)